requirements.txt
numpy~=2.0.2
opencv-python~=4.11.0.86
requests~=2.32.3
pillow~=11.1.0
selenium~=4.29.0
pynput~=1.8.1
stable-baselines3~=2.5.0
ultralytics~=8.3.93